[IR][SPARK] Reorganized spark optimizations API.
- Moved `spark.SparkSpecializeSupport` to `opt.SparkSpecializeOps`.
- Installed top-level API in trait `opt.SparkOptimizations`.
